refactor: orders wave 6 slice A — validation splits and 350 ratchet (#4246)
Split order_engine_validation phase runners and work_order_validator
material/tech helpers so grandfathered modules clear the wave-6 ceiling.
Ratchet orders lib physical-line limit to 350 with shrink-only entries for
near-cap suggestion/application files deferred to Slice C.

Refs #4246
